NautobotCleaner/pyproject.toml
[tool.poetry]
name = "nautbotcleaner"
version = "0.1.0"
description = ""
authors = ["Angelo Poggi <angelo.poggi@opti9tech.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pynautobot = "^1.0.4"
python-dotenv = "^0.20.0"
napalm = "^3.4.0"
gevent = "^21.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
